component관리
