+++ frontend/src/components/ShopList.jsx
@@ -5,6 +5,8 @@ function ShopList() {
     const [items, setItems] = useState([]);
     const [remainTime, setRemainTime] = useState('');
     const [loading, setLoading] = useState(true);
+    const [error, setError] = useState(false);
+
 
     // 시간 문자열을 초(int)로 변환
     const timeToSeconds = (timeStr) => {
@@ -23,12 +25,15 @@ function ShopList() {
     const fetchShop = async () => {
         try {
             setLoading(true);
+            setError(false); // 초기화
             const res = await axios.get('/api/shop');
             setItems(res.data.items);
             setRemainTime(res.data.remainTime);
-            setLoading(false);
         } catch (err) {
             console.error('마리샵 로딩 실패:', err);
+            setError(true); // 실패 표시
+            setItems([]);   // 빈 배열
+        } finally {
             setLoading(false);
         }
     };
@@ -56,16 +61,7 @@ function ShopList() {
         return () => clearInterval(timer);
     }, [remainTime]);
 
-    // 테스트용 더미 데이터 (API 연결이 안될 경우)
-    const dummyItems = [
-        { title: '여름 코스튬 세트', price: '2400', isNew: true, image: '' },
-        { title: '캐릭터 슬롯 확장권', price: '800', image: '' },
-        { title: '프리미엄 아바타 상자', price: '1950', isNew: true, secondPrice: '2340', image: '' },
-        { title: '성장 지원 패키지', price: '1200', isLimited: true, image: '' },
-        { title: '펫 상자', price: '1000', image: '' }
-    ];
 
-    const displayItems = items.length > 0 ? items : dummyItems;
 
     return (
         <div className="mb-8">
@@ -73,6 +69,14 @@ function ShopList() {
                 <div className="flex justify-center items-center h-20">
                     <div className="animate-spin rounded-full h-6 w-6 border-t-2 border-b-2 border-gray-900 dark:border-white"></div>
                 </div>
+            ) : error ? (
+                <p className="text-center text-red-500">
+                    마리샵 API 연결에 실패했습니다. 나중에 다시 시도해주세요.
+                </p>
+            ) : items.length === 0 ? (
+                <p className="text-center text-gray-500 dark:text-gray-400">
+                    현재 입고된 상품이 없습니다.
+                </p>
             ) : (
                 <>
                     {remainTime && (
@@ -80,12 +84,12 @@ function ShopList() {
                             <svg xmlns="http://www.w3.org/2000/svg" className="h-5 w-5 mr-1" viewBox="0 0 20 20" fill="currentColor">
                                 <path fillRule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zm1-12a1 1 0 10-2 0v4a1 1 0 00.293.707l2.828 2.829a1 1 0 101.415-1.415L11 9.586V6z" clipRule="evenodd" />
                             </svg>
-                            다음 입고까지: {remainTime}
+                            새 상품 입고까지: {remainTime}
                         </div>
                     )}
 
                     <div className="space-y-3">
-                        {displayItems.map((item, index) => (
+                        {items.map((item, index) => (
                             <div key={index} className="bg-white dark:bg-gray-800 rounded-lg border border-gray-200 dark:border-gray-700 overflow-hidden shadow-sm">
                                 <div className="flex items-center p-4">
                                     <div className="flex-shrink-0 w-16 h-16 bg-gray-200 dark:bg-gray-700 rounded">
@@ -95,8 +99,6 @@ function ShopList() {
                                     <div className="ml-4 flex-grow">
                                         <div className="flex items-center">
                                             <h3 className="text-lg font-medium text-gray-900 dark:text-white">{item.title}</h3>
-
-                                            {/* 태그 표시 */}
                                             {item.isNew && (
                                                 <span className="ml-2 px-2 py-1 text-xs font-semibold rounded bg-blue-500 text-white">신규</span>
                                             )}
@@ -107,15 +109,15 @@ function ShopList() {
 
                                         <div className="flex items-center mt-1 space-x-2">
                                             <div className="flex items-center space-x-1">
-                                                <span className="text-lg font-semibold text-purple-500">
-                                                    💎 {item.price}
-                                                </span>
+                                    <span className="text-lg font-semibold text-purple-500">
+                                        💎 {item.price}
+                                    </span>
                                             </div>
 
                                             {item.originalPrice && (
                                                 <span className="text-sm text-gray-400 line-through">
-                                                    {item.originalPrice}
-                                                </span>
+                                        {item.originalPrice}
+                                    </span>
                                             )}
                                         </div>
                                     </div>
@@ -125,6 +127,7 @@ function ShopList() {
                     </div>
                 </>
             )}
+
         </div>
     );
 }
